requirements.txt
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.6.3
attrs==25.3.0
bcrypt==3.2.0
beautifulsoup4==4.13.4
blinker==1.7.0
cachetools==4.2.2
certifi==2025.7.14
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
ecdsa==0.18.0
fastapi==0.109.1
filelock==3.14.0
Flask==3.0.2
h11==0.14.0
httptools==0.6.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
nltk==3.8.1
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
outcome==1.3.0.post0
packaging==25.0
passlib==1.7.4
pillow==10.3.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
PySocks==1.7.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.7
python-telegram-bot==13.15
pytz==2025.2
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rsa==4.9
selenium==4.34.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.7
starlette==0.35.1
sympy==1.12
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tornado==6.1
tqdm==4.66.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.27.0.post1
watchfiles==0.21.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
